requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blinker==1.8.2
boto3==1.34.143
botocore==1.34.143
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.8
exceptiongroup==1.2.1
fastavro==1.9.5
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.1
parameterized==0.9.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.32.0.20240712
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
